+++ examples/example-orbital.c
@@ -5,7 +5,15 @@
  * @author Attila Kovacs
  *
  *  Example file for using the SuperNOVAS C/C++ library for determining positions for
- *  nearby (non-high-z) sidereal sources, such as a star.
+ *  Solar-system objects define through a set of orbital parameters.
+ *
+ *  For example, the IAU Minor Planet Center (MPC) publishes current orbital
+ *  parameters for known asteroids, comets, and near-Earth objects. While orbitals are
+ *  not super precise in general, they can provide sufficienly accurate positions on
+ *  the arcsecond level (or below), and may be the best/only source of position data
+ *  for newly discovered objects.
+ *
+ *  See https://minorplanetcenter.net/data
  *
  *  Link with
  *
@@ -32,7 +40,7 @@
 
 int main() {
   // SuperNOVAS aariables used for the calculations ------------------------->
-  cat_entry star = CAT_ENTRY_INIT;  // catalog information about a sidereal source
+  novas_orbital orbit = NOVAS_ORBIT_INIT;     // Orbital parameters
   object source;                    // a celestial object: sidereal, planet, ephemeris or orbital source
   observer obs;                     // observer location
   novas_timespec obs_time;          // astrometric time of observation
@@ -51,36 +59,54 @@ int main() {
   novas_debug(NOVAS_DEBUG_ON);
 
 
+  // Orbitals assume Keplerian motion, and are never going to be accurate much below the
+  // tens of arcsec level even for the most current MPC orbits. Orbitals for planetary
+  // satellites are even less precise. So, with orbitals, there is no point on pressing
+  // for ultra-high (sub-uas level) accuracy...
+  accuracy = NOVAS_REDUCED_ACCURACY;      // mas-level precision, typically
+
+
   // -------------------------------------------------------------------------
   // Define a sidereal source
 
-  // Let's assume we have B1950 (FK4) coordinates...
-  // 16h26m20.1918s, -26d19m23.138s (B1950), proper motion -12.11, -23.30 mas/year,
-  // parallax 5.89 mas, radial velocity -3.4 km/s.
-  if(make_cat_entry("Antares", "FK4", 1, 16.43894213, -26.323094, -12.11, -23.30, 5.89, -3.4, &star) != 0) {
-    fprintf(stderr, "ERROR! defining cat_entry.\n");
-    return 1;
-  }
+  // Orbital Parameters for the asteroid Ceres from the Minor Planet Center
+  // (MPC) at JD 2460600.5
+  orbit.jd_tdb = 2460600.5;   // [day] TDB date
+  orbit.a = 2.7666197;        // [AU]
+  orbit.e = 0.079184;
+  orbit.i = 10.5879;          // [deg]
+  orbit.omega = 73.28579;     // [deg]
+  orbit.Omega = 80.25414;     // [deg]
+  orbit.M0 = 145.84905;       // [deg]
+  orbit.n = 0.21418047;       // [deg/day]
 
-  // First change the catalog coordinates (in place) to the J2000 (FK5) system...
-  if(transform_cat(CHANGE_EPOCH, NOVAS_JD_B1950, &star, NOVAS_JD_J2000, "FK5", &star) != 0) {
-    fprintf(stderr, "ERROR! converting B1950 catalog coordinates to J2000.\n");
-    return 1;
-  }
-
-  // Then convert J2000 coordinates to ICRS (also in place). Here the dates don't matter...
-  if(transform_cat(CHANGE_J2000_TO_ICRS, 0.0, &star, 0.0, "ICRS", &star) != 0) {
-    fprintf(stderr, "ERROR! converting J2000 catalog coordinates to ICRS.\n");
-    return 1;
-  }
+  // Define Ceres as the observed object (we can use whatever ID numbering
+  // system here, since it's irrelevant to SuperNOVAS in this context).
+  make_orbital_object("Ceres", 2000001, &orbit, &source);
 
 
-  // -------------------------------------------------------------------------
-  // Wrap the sidereal souce into an object structure...
-  if(make_cat_object(&star, &source) != 0) {
-    fprintf(stderr, "ERROR! configuring observed object\n");
-    return 1;
-  }
+  // ... Or, you could define orbitals for a satellite instead:
+  /*
+  // E.g. Callisto's orbital parameters from JPL Horizons
+  // https://ssd.jpl.nasa.gov/sats/elem/sep.html
+  // 1882700. 0.007 43.8  87.4  0.3 309.1 16.690440 277.921 577.264 268.7 64.8
+  orbit.system.center = NOVAS_JUPITER;
+  novas_set_orbsys_pole(NOVAS_GCRS, 268.7 / 15.0, 64.8, &orbit->system);
+
+  orbit.jd_tdb = NOVAS_JD_J2000;
+  orbit.a = 1882700.0 * 1e3 / AU;
+  orbit.e = 0.007;
+  orbit.omega = 43.8;
+  orbit.M0 = 87.4;
+  orbit.i = 0.3;
+  orbit.Omega = 309.1;
+  orbit.n = TWOPI / 16.690440;
+  orbit.apsis_period = 277.921 * 365.25;
+  orbit.node_period = 577.264 * 365.25;
+
+  // Set Callisto as the observed object
+  make_orbital_object("Callisto", 501, &orbit, &source);
+  */
 
 
   // -------------------------------------------------------------------------
@@ -102,38 +128,23 @@ int main() {
   clock_gettime(CLOCK_REALTIME, &unix_time);
 
   // Set the time of observation to the precise UTC-based UNIX time
+  // (We can set astromtric time using an other time measure also...)
   if(novas_set_unix_time(unix_time.tv_sec, unix_time.tv_nsec, LEAP_SECONDS, DUT1, &obs_time) != 0) {
     fprintf(stderr, "ERROR! failed to set time of observation.\n");
     return 1;
   }
 
-  // ... Or you could set a time explicily in any known timescale.
-  /*
-  // Let's set a TDB-based time for the start of the J2000 epoch exactly...
-  if(novas_set_time(NOVAS_TDB, NOVAS_JD_J2000, LEAP_SECONDS, DUT1, &obs_time) != 0) {
-    fprintf(stderr, "ERROR! failed to set time of observation.\n");
-    return 1;
-  }
-  */
-
 
   // -------------------------------------------------------------------------
   // You might want to set a provider for precise planet positions so we might
-  // calculate Earth, Sun and major planet positions accurately. If an planet
-  // provider is configured, we can unlock the ultimate (sub-uas) accuracy of
-  // SuperNOVAS.
+  // calculate Earth, Sun and major planet positions accurately. It is needed
+  // if you have orbitals defined around a major planet.
   //
   // There are many ways to set a provider of planet positions. For example,
   // you may use CALCEPH:
   //
   // t_calcephbin *planets = calceph_open("path/to/de440s.bsp");
   // novas_use_calceph(planets);
-  //
-  // accuracy = NOVAS_FULL_ACCURACY;      // sub-uas precision
-
-  // Without a planet provider, we are stuck with reduced (mas) precisions
-  // only...
-  accuracy = NOVAS_REDUCED_ACCURACY;      // mas-level precision, typically
 
 
   // -------------------------------------------------------------------------
@@ -144,7 +155,6 @@ int main() {
     fprintf(stderr, "ERROR! failed to define observing frame.\n");
     return 1;
   }
-
 
   // -------------------------------------------------------------------------
   // Calculate the precise apparent position (here in CIRS coordinates)